Conflicts:
	app/server/services/joinChat.coffee
	app/server/services/kickUser.coffee
	data/factory.coffee
	test/server/specs/service.acceptChat.coffee
	test/server/specs/service.populateVisitorAcpData.coffee
	test/server/specs/service.query.coffee

+++ app/server/services/kickUser.coffee
@@ -18,6 +18,9 @@ module.exports =
       async.map chatSessions, getRole, (err, chatSessions) ->
         [visitorChatSession] = chatSessions.filter (s) -> s.role is 'Visitor'
 
+        unless visitorChatSession
+          return done()
+
         # remove the visitor from the chat
         visitorChatSession.session.delete (err) ->
           config.log.error 'Error deleting session in kickUser', {error: err, chatId: chatId} if err
@@ -26,4 +29,4 @@ module.exports =
           notify = pulsar.channel chatId
           notify.emit 'chatEnded'
 
-          done null, null
+          done()
